+++ module/batch-action/src/Application/DependencyInjection/ErgonodeBatchActionExtension.php
@@ -63,19 +63,6 @@ class ErgonodeBatchActionExtension extends Extension implements PrependExtension
 
     private function prependMessenger(ContainerBuilder $container): void
     {
-        $configs = $container->getExtensionConfig($this->getAlias());
-        $configuration = $this->getConfiguration($configs, $container);
-        $config = $this->processConfiguration($configuration, $configs);
-
-        if (!$this->isConfigEnabled($container, $config['messenger'])) {
-            return;
-        }
-
-        $container->setParameter(
-            'ergonode.batch_action.messenger_transport_name',
-            $config['messenger']['transport_name'],
-        );
-
         $loader = new YamlFileLoader($container, new FileLocator(__DIR__.'/../../Resources/config'));
 
         $loader->load('messenger.yaml');