+++ module/batch-action/src/Domain/Repository/BatchActionRepositoryInterface.php
@@ -25,4 +25,6 @@ interface BatchActionRepositoryInterface
      * @param BatchActionMessage[] $messages
      */
     public function markEntry(BatchActionId $id, AggregateId $resourceId, array $messages): void;
+
+    public function endBatchAction(BatchActionId $id): void;
 }