+++ module/batch-action/src/Infrastructure/Handler/ProcessBatchActionEntryCommandHandler.php
@@ -11,6 +11,8 @@ namespace Ergonode\BatchAction\Infrastructure\Handler;
 use Ergonode\BatchAction\Domain\Command\ProcessBatchActionEntryCommand;
 use Ergonode\BatchAction\Domain\Repository\BatchActionRepositoryInterface;
 use Ergonode\BatchAction\Infrastructure\Provider\BatchActionProcessorProvider;
+use Webmozart\Assert\Assert;
+use Ergonode\BatchAction\Domain\Entity\BatchAction;
 
 class ProcessBatchActionEntryCommandHandler
 {
@@ -18,21 +20,27 @@ class ProcessBatchActionEntryCommandHandler
 
     private BatchActionRepositoryInterface $repository;
 
-    public function __construct(BatchActionProcessorProvider $provider, BatchActionRepositoryInterface $repository)
-    {
+    public function __construct(
+        BatchActionProcessorProvider $provider,
+        BatchActionRepositoryInterface $repository
+    ) {
         $this->provider = $provider;
         $this->repository = $repository;
     }
 
-    public function __invoke(ProcessBatchActionEntryCommand $command): void
+    public function __invoke(ProcessBatchActionEntryCommand $command): array
     {
-        $id = $command->getId();
-        $type = $command->getType();
+        $batchAction = $this->repository->load($command->getId());
+
+        Assert::isInstanceOf($batchAction, BatchAction::class);
+
+        $id = $batchAction->getId();
+        $type = $batchAction->getType();
         $resourceId = $command->getResourceId();
-        $payload = $command->getPayload();
+        $payload = $batchAction->getPayload();
 
         $processor = $this->provider->provide($type);
-        $message = $processor->process($id, $resourceId, $payload);
-        $this->repository->markEntry($id, $resourceId, $message);
+
+        return $processor->process($id, $resourceId, $payload);
     }
 }