+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## CHANGELOG FOR 1.2.x
 #### 1.2.0
+- refactor [#1556](https://github.com/ergonode/backend/issues/1556) Batch Action - refactoring of resume process (rprzedzik)
 - refactor [#1552](https://github.com/ergonode/backend/issues/1552) Add tag support for ErgonodeProcessorStepInterface (rprzedzik)
 - refactor [#1542](https://github.com/ergonode/backend/issues/1542) Opening the file export extension to other modules (wfajczyk)
 - bugfix [#1538](https://github.com/ergonode/backend/issues/1538) Bug when denormalize int value as float type in json (rprzedzik)